% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_summary.R
\name{basic_summary}
\alias{basic_summary}
\title{Basic summary of dataset}
\usage{
basic_summary(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
summary output
}
\description{
Basic summary of dataset
}
