% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Remove missing values from a dataset}
\usage{
clean_data(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
cleaned data.frame
}
\description{
Remove missing values from a dataset
}
