% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot.R
\name{quick_plot}
\alias{quick_plot}
\title{Quick scatter plot}
\usage{
quick_plot(df, x, y)
}
\arguments{
\item{df}{data.frame}

\item{x}{column name (string)}

\item{y}{column name (string)}
}
\value{
ggplot object
}
\description{
Quick scatter plot
}
